add web client and weather API consume

+++ src/main/java/infnet/edu/springreactiveproject/repository/ProductRepository.java
@@ -4,6 +4,6 @@ import infnet.edu.springreactiveproject.model.Product;
 import org.springframework.data.r2dbc.repository.R2dbcRepository;
 import reactor.core.publisher.Flux;
 
-public interface ProductRepository extends R2dbcRepository<Product, Long> {
+public interface ProductRepository extends R2dbcRepository<Product, Integer> {
     Flux<Product> findByName(String name);
 }
